stateless session cookie
Implementation is based on http://www.cl.cam.ac.uk/~sjm217/papers/protocols08cookies.pdf . The server-side secret that used for hmac calculation for the cookie is derived from a file outside the backup area and per-user cookie salt. The latter allows to implement logout-all functionality for the user as the change in the salt invalidates all login cookies for the user without affecting others.