removal of title page
The link to add the new text is moved to the list page making the title page unnecessary. The link to delete the page is moved to the edit screen and the edit is now activated from the text itself.

Another GUI change is turning the login form into one-liner to safe the space.

Internally I split the logic/presentation for list/show so all HTML generation are now is done in database.php file.