wordpress update and config/content outside wp dir
This updates to WordPress 4.2.3 and moves wp-config to the parent directory of wordpress where it also sets WP_CONTENT_DIR. This way future upgrades of wordpress require just unpacking the release archive and committing the changes. There is no need to worry about preserving wp-content and wp-config.php in the directory.