direct mount of php config files
Passing the real location of the config files over Fastcgi parameters just complicates the setup. So I replaced that with a direct mount of config files into /www/soft tree leading to simpler and more robust setup.

In particular, to update fluxbb it is sufficient to remove the current fluxbb directory and extract the distribution into it. This is similar to what was possible with wordpress.